Text Version
Decided to make a text version of the game before doing the GUI stuff.
created TextVersion.java, added to Player.java. When the program is run,
it prints out a 9x9 array of Squares with only the castle in the center,
and gives the player one random domino. The player then specifies the XY
coords and the heading, and if it is a legal placement, it updates the
9x9 array with the newly placed domino.