Out Of Bounds
Players are no longer able to place pieces that would make their Kingdom
larger than a 5x5 grid. Each time you place a piece, it checks for the
min and max X,Y values of placed tiles in 9x9 grid, and then uses those
values to mark squares that would be outside a 5x5 grid as OUTOFBOUNDS,
denoted by an X, when the player board is displayed.