Scoring
A score is now correctly calculated for a completed player board.
Added to Player class: 
the findAZone method, that looks a square at a given x,y value on the
player board, then recursively finds adjacent tiles of the same type and
adds them to a list representing that "Zone", and marks the tiles and
being scored.

The scoreAZone method, which has an x,y parameter, calls the findAZone
method at an x,y value on the player board, and calculates that zone's
score based on the number of Squares in  the "theZone" list, multiplying
that number by the number of crowns in that zone. It adds that zone
score to the total board score, then clears the "theZone" list. 

The scoreTheBoard method, which loops the scoreAZone method for every
x,y value on the player board.