Refactoring
Created DominoPlacement class, which represents a placement of a domino
(x,y, the domino toBePlaced, heading). All of the methods in TextVersion
that were taking those four values as parameters, and now taking a
DominoPlacement object instead. 
Extracted out of the placeTheDomino method two new methods:
specifyPlacement gathers the X,Y and heading values from the user into a
DominoPlacement object, and makePlacement places the Domino according
the the values of the DominoPlacement Object returned by
specifyPlacement