Multiple Players, Turns, PlayArea
The game now supports multiple players, which the user specifies on app
startup. Created Turn class, which is the display of dominoes to be
chosen at the beginning of each round. Created PlayArea class, which
represents a Player's area where they will place dominos (a 10x10 grid
of Square objects.) The PlayArea class now has the methods for scoring a
complete board. Moved all of the methods related to placing a domino
into the Player class, and renamed the TextVersion class to TurnLooper,
which is now only responsible for starting the game(getting the player
count), gathering the XY coordinates of a domino to be placed, and
looping through 12 turns. At this point, the text version of the game is
complete, minus some error handling.